[refactor] post 패키지에 category 패키지 넣기
