+++ backend/src/main/java/com/NBE_4_5_2/Team5/domain/post/category/controller/CategoryController.java
@@ -1,7 +1,7 @@
-package com.NBE_4_5_2.Team5.domain.category.controller;
+package com.NBE_4_5_2.Team5.domain.post.category.controller;
 
-import com.NBE_4_5_2.Team5.domain.category.entity.Category;
-import com.NBE_4_5_2.Team5.domain.category.service.CategoryService;
+import com.NBE_4_5_2.Team5.domain.post.category.entity.Category;
+import com.NBE_4_5_2.Team5.domain.post.category.service.CategoryService;
 import com.NBE_4_5_2.Team5.global.dto.RsData;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
