+++ backend/src/main/java/com/NBE_4_5_2/Team5/domain/post/category/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
-package com.NBE_4_5_2.Team5.domain.category.repository;
+package com.NBE_4_5_2.Team5.domain.post.category.repository;
 
-import com.NBE_4_5_2.Team5.domain.category.entity.Category;
+import com.NBE_4_5_2.Team5.domain.post.category.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
