+++ backend/src/main/java/com/NBE_4_5_2/Team5/domain/post/category/service/CategoryService.java
@@ -1,7 +1,7 @@
-package com.NBE_4_5_2.Team5.domain.category.service;
+package com.NBE_4_5_2.Team5.domain.post.category.service;
 
-import com.NBE_4_5_2.Team5.domain.category.entity.Category;
-import com.NBE_4_5_2.Team5.domain.category.repository.CategoryRepository;
+import com.NBE_4_5_2.Team5.domain.post.category.entity.Category;
+import com.NBE_4_5_2.Team5.domain.post.category.repository.CategoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
