+++ backend/src/main/java/com/NBE_4_5_2/Team5/domain/post/post/entity/ProductCategory.java
@@ -1,6 +1,6 @@
 package com.NBE_4_5_2.Team5.domain.post.post.entity;
 
-import com.NBE_4_5_2.Team5.domain.category.entity.Category;
+import com.NBE_4_5_2.Team5.domain.post.category.entity.Category;
 import jakarta.persistence.*;
 import lombok.*;
